// libraries/visualiser-library/src/Visualiser/Functions/ThreeJsUtils/Gltf/types/KHR_texture_transform.textureInfo.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * glTF extension that enables shifting and scaling UV coordinates on a per-texture basis
 */
export type KHRTextureTransformTextureInfoExtension = KHRTextureTransformTextureInfoExtension1 &
  KHRTextureTransformTextureInfoExtension2;
export type KHRTextureTransformTextureInfoExtension1 = GlTFProperty & {
  /**
   * The offset of the UV coordinate origin as a factor of the texture dimensions.
   */
  offset?: [
    {
      [k: string]: unknown | undefined;
    } & number,
    {
      [k: string]: unknown | undefined;
    } & number
  ];
  /**
   * Rotate the UVs by this many radians counter-clockwise around the origin.
   */
  rotation?: number;
  /**
   * The scale factor applied to the components of the UV coordinates.
   */
  scale?: [
    {
      [k: string]: unknown | undefined;
    } & number,
    {
      [k: string]: unknown | undefined;
    } & number
  ];
  /**
   * Overrides the textureInfo texCoord value if supplied, and if this extension is supported.
   */
  texCoord?: number;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown | undefined;
};

export interface GlTFProperty {
  extensions?: Extension;
  extras?: Extras;
  [k: string]: unknown | undefined;
}
/**
 * Dictionary object with extension-specific objects.
 */
export interface Extension {
  [k: string]:
    | {
        [k: string]: unknown | undefined;
      }
    | undefined;
}
/**
 * Application-specific data.
 */
export interface Extras {
  [k: string]: unknown | undefined;
}
export interface KHRTextureTransformTextureInfoExtension2 {
  /**
   * The offset of the UV coordinate origin as a factor of the texture dimensions.
   */
  offset?: [
    {
      [k: string]: unknown | undefined;
    } & number,
    {
      [k: string]: unknown | undefined;
    } & number
  ];
  /**
   * Rotate the UVs by this many radians counter-clockwise around the origin.
   */
  rotation?: number;
  /**
   * The scale factor applied to the components of the UV coordinates.
   */
  scale?: [
    {
      [k: string]: unknown | undefined;
    } & number,
    {
      [k: string]: unknown | undefined;
    } & number
  ];
  /**
   * Overrides the textureInfo texCoord value if supplied, and if this extension is supported.
   */
  texCoord?: number;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown | undefined;
}